1.20.1 - Implemented special creative tab items 1.20.1 Port - Fixed item descriptions 1.20.1 Port - Missing assets 1.20.1 Port - Fixed resources 1.20.1 Port - Compilation fixes

+++ src/main/java/com/buuz135/industrial/proxy/client/sound/TickeableSound.java
@@ -102,7 +102,7 @@ public class TickeableSound extends AbstractSoundInstance implements TickableSou
 
     @Override
     public void tick() {
-        double distance = Minecraft.getInstance().player.blockPosition().distManhattan(new BlockPos(this.x, this.y, this.z));
+        double distance = Minecraft.getInstance().player.blockPosition().distManhattan(new BlockPos((int) this.x, (int) this.y, (int) this.z));
         if (distance > this.distance) {
             this.volume = 0;
         } else {
